backend/requirements.txt
# MangaHub AI Backend - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP Client (for translation API)
httpx>=0.26.0

# Image Processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# Manga OCR (specialized for Japanese manga)
manga-ocr>=0.1.11
transformers>=4.35.0
torch>=2.0.0

# Speech Bubble Detection (YOLOv8)
ultralytics>=8.0.0
huggingface_hub>=0.20.0

# OCR - PaddleOCR (optional, for other languages)
# paddlepaddle>=2.6.0
# paddleocr>=2.7.0

# Development
python-dotenv>=1.0.0
